backend/ml/requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
xgboost==2.1.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
joblib==1.4.2
python-dotenv==1.0.1
